requirements.txt
opencv-python-headless
cvlib
streamlit
Pillow
